Ideal release process for Windows detached signing
This is an ideal version of what the release process should look like,
making it more consistent with the OS X process. Some of the changes
described here would need to be made in the descriptors, which is somewhat
beyond what I would feel comfortable doing, not really understanding the signature process in depth.

[skip ci]

Github-Pull: #6354
Rebased-From: 0710016f8b57035c8edb342183e910dc6653e778