Further Enhancement for Priority Queuing
Here we just bypass putting the return xthinblock into the
INV gedata queue.  We just send it directly back with a call
to SendXthinBlock()

Also, Both the XTHIN and BLOOM service need to be turned on for Xthins to work

We always needed both of these services however we were only checking for the XTHIN
service when in fact we need both to be on.