Edge case Fix for not allowing two addnode connections to be made to the same peer
During startup it is still possible for two addnode nodes to connect both in
and outbound.  Putting a 15 second delay at startup allows the remote node
to connect before we make the attempt.  The real problem lies in the fact that
we don't update the list of possible -addnode nodes every time we make a connection
but that would be more invasive a fix.