New configuration parameter: maxoutconnections
Remove the hard coded maximum number of outbound connections.
Add a new parameter to let the operator to set the number of
outbound connections.

newoutconnections default value is 8 and it can't be greater
that maxconnections, if this is the case it will be automatic
decreased to maxconnections. In such cases nodes will establish
only outbound connections.