do not allow HTTP timeouts to happen in excessive.py
We set the http keepalive timeout to zero to prevent a timeout from
happneing while we're waiting for a process to finish and then
we end up trying to request from a connection that no longer exists.
Instead we timout the connections right away so that each time a new
fresh connection is aquired.

This is something that has been coming up lately in very lately in long
running tests.