Comptool fix for the requestmanager re-request interval
When a block doesn't show up right away the request manager
will wait 5 seconds (on mainnet it is 30 seconds) before
re-requesting a block.  If that should happen during a test
we need to check whether the block was successfully re-requested
and if so mark the test as passed.  Previously we were waiting
for the 5 seconds but then returning with a False if the correct
block showed up when it should be True.