Update UI banned list
-When a ban/unban happens internally (not through RPC) we need
to refresh the ban list in the UI.  Previous to this we
were only doing it if the ban was set through RPC only.

-Add 5 second Qtimer for automatically updating the banned list

When a ban comes to an end it will get deleted from the UI banned
list automatically rather than having to wait for another node
to be banned before the list would update.

-Keep use of SweepBanned() consistent for UI banned list

Dumping the banned list to disk every time the banned list is updated
is not necessary.  It gets dumped to disk every 15 mins and also when
the node shuts down.  Using sweepbanned is also consistent now with
the rest of the calls to bannning and unbanning of nodes.

-Several places in RPC calls for banning and unbanning
with redundant code.

-Make all calls to uiInterface.BannnedListChanged in src/net.cpp
This makes the code consistent for future use

-Add sweepbanned to listbanned