Disconnect and ban a node if they fail to provide the inital HEADERS
At the beginning of an initial sync it is possible for an attack node
to prevent the initial download of HEADERS required for the initial
sync to proceed. In the event that we do not receive the initial batch
of headers from a node then disconnect them safely with the fDisconnect
flag and then give them a 4 hour ban.