This change affects the MaxGeneratedBlock, ExcessiveBlockSize & ExcessiveAcceptDepth qt text field widgets in the bitcoin unlimited settings dialog.
Before this change, the user could enter any junk text into these feilds and the dialog would accept that junk. When the user hit 'ok' then the settings would silently fail & the user would not be informed of the error or failure.

This change places QT integer validators on these text fields so that only a valid range of digits and integer value can be entered thus removing the 'silent fail' issue. This change also cleans up and simplifies the validation code for these fields. The changes are more consistant with the implementation of text fields elsewhere in the GUI (most of the other dialogs use QT Validators on textfields).

This change improves the user experience and removes the 'silent fail' that could happen when configuring these key fields.