Refactor miner.cpp code slightly
Break out the logic to test whether an incremental increase in block size and
sigop count is good.  No change in logic or functionality.
This will make it easier to review and merge the CPFP changes where txs are added
as a "package" not as single txs whilst retaining the BU block validity logic.