Fix opt-in RBF reliance on compiler integer size
Opt-in RBF was using std::numeric_limits<unsigned int>::max() for
determining whether a TXN is RBF- able or not, thus relying on an
architecture/compiler detail, which would fail e.g. for ILP64 data types.

This commit replaces the above with the SEQUENCE_FINAL protocol constant.

Note that opt-in RBF is disabled in BU with an if(0) { ... }, so this is
to fix the code in case it ever gets re-enabled or used in any way.

It is a port of this pull request submitted to Core:

https://github.com/bitcoin/bitcoin/pull/10172