[Refactor] Eliminate using namespace std in main.cpp
"using namespace std" is a code smell that should be eliminated.
1. Wrote python script to update standard library types to use std::
instead of relying on "using namespace std".
2. Ran script on main.cpp and eliminated use of "using namespace std" in
this file.

NOTE: The python script is still very preliminary.  It includes less
than 1% of the C++ standard library keywords and due to liminatations in
native lookbehind requring fixed width search is not very robust.
Current form is sufficient to convert main.cpp, but runs on other files
will likely require additional updates to the python script.